requirements.txt
flask
google-generativeai
mediapipe
numpy
pandas
opencv-python
python-dotenv
requests
